+++ rsass-macros/src/lib.rs
@@ -70,7 +70,7 @@ pub fn include_scss(tokens: TokenStream) -> TokenStream {
                 .to_token_stream()
                 .into()
         }
-        Err(err) => Error::new(Span::call_site(), format!("{err:?}"))
+        Err(err) => Error::new(Span::call_site(), format!("{err}"))
             .into_compile_error()
             .into(),
     }